cleaned up encoders.py and encoders_base.py
